# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_track_v1_recordings__recording_id__tracks_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "kind": {
            "enum": [
              "audio",
              "midi"
            ],
            "title": "Kind",
            "type": "string"
          },
          "offsetMs": {
            "default": 0,
            "title": "Offsetms",
            "type": "integer"
          }
        },
        "required": [
          "file",
          "kind"
        ],
        "title": "Body_upload_track_v1_recordings__recording_id__tracks_post",
        "type": "object"
      },
...
      "Page_RecordingSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecordingSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RecordingSummary]",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 (problem+json) error body.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "RecordingCreate": {
        "properties": {
          "capturedAt": {
            "format": "date-time",
            "title": "Capturedat",
            "type": "string"
          },
          "durationMs": {
            "anyOf": [
              {
                "maximum": 86400000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sessionId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "subjectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectid"
          },
          "subjectKind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectkind"
          }
        },
        "required": [
          "capturedAt"
        ],
        "title": "RecordingCreate",
        "type": "object"
      },
      "RecordingRead": {
        "properties": {
          "capturedAt": {
            "format": "date-time",
            "title": "Capturedat",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "durationMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sessionId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "subjectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectid"
          },
          "subjectKind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectkind"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/RecordingTrackRead"
            },
            "title": "Tracks",
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectKind",
          "subjectId",
          "sessionId",
          "capturedAt",
          "durationMs",
          "notes",
          "createdAt",
          "updatedAt",
          "tracks"
        ],
        "title": "RecordingRead",
        "type": "object"
      },
      "RecordingSummary": {
        "description": "The list-view shape \u2014 no tracks, so `GET /v1/recordings` stays one\nquery per page rather than N+1 (see `RecordingRead` for the\nsingle-recording shape that does carry them) \u2014 the same split\n`IdeaSummary`/`IdeaRead` use in `app/schemas/idea.py`, for the same\nreason.",
        "properties": {
          "capturedAt": {
            "format": "date-time",
            "title": "Capturedat",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "durationMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sessionId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "subjectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectid"
          },
          "subjectKind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectkind"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectKind",
          "subjectId",
          "sessionId",
          "capturedAt",
          "durationMs",
          "notes",
          "createdAt",
          "updatedAt"
        ],
        "title": "RecordingSummary",
        "type": "object"
      },
      "RecordingTrackRead": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "audio",
              "midi"
            ],
            "title": "Kind",
            "type": "string"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "offsetMs": {
            "title": "Offsetms",
            "type": "integer"
          },
          "recordingId": {
            "format": "uuid",
            "title": "Recordingid",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "storageKey": {
            "title": "Storagekey",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "recordingId",
          "kind",
          "storageKey",
          "mime",
          "bytes",
          "sha256",
          "offsetMs",
          "createdAt",
          "updatedAt"
        ],
        "title": "RecordingTrackRead",
        "type": "object"
      },
      "RecordingUpdate": {
        "properties": {
          "durationMs": {
            "anyOf": [
              {
                "maximum": 86400000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "RecordingUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Soundings API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/recordings": {
      "get": {
        "operationId": "list_recordings_v1_recordings_get",
        "parameters": [
          {
            "in": "query",
            "name": "subjectKind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subjectkind"
            }
          },
          {
            "in": "query",
            "name": "subjectId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subjectid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RecordingSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          }
        },
        "summary": "List Recordings",
        "tags": [
          "recordings"
        ]
      },
      "post": {
        "operationId": "create_recording_v1_recordings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          }
        },
        "summary": "Create Recording",
        "tags": [
          "recordings"
        ]
      }
    },
    "/v1/recordings/{recording_id}": {
      "delete": {
        "operationId": "delete_recording_v1_recordings__recording_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          }
        },
        "summary": "Delete Recording",
        "tags": [
          "recordings"
        ]
      },
      "get": {
        "operationId": "get_recording_v1_recordings__recording_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          }
        },
        "summary": "Get Recording",
        "tags": [
          "recordings"
        ]
      },
      "patch": {
        "operationId": "update_recording_v1_recordings__recording_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          }
        },
        "summary": "Update Recording",
        "tags": [
          "recordings"
        ]
      }
    },
    "/v1/recordings/{recording_id}/tracks": {
      "post": {
        "operationId": "upload_track_v1_recordings__recording_id__tracks_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "content-length",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content-Length"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_track_v1_recordings__recording_id__tracks_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingTrackRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          }
        },
        "summary": "Upload Track",
        "tags": [
          "recording-tracks"
        ]
      }
    },
    "/v1/recordings/{recording_id}/tracks/{track_id}/content": {
      "get": {
        "operationId": "download_track_v1_recordings__recording_id__tracks__track_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Track Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (problem+json)"
          }
        },
        "summary": "Download Track",
        "tags": [
          "recording-tracks"
        ]
      }
    },
...
  }
}